requirements.txt
streamlit==1.36.0
PyPDF2==3.0.1
python-docx==1.1.0
openai==1.35.1
google-generativeai==0.7.0